Solution to problem 2

+++ e002.pyx
@@ -4,9 +4,9 @@
 
 cdef extern from "math.h":
     double sqrt(double x)
-    long lround(double x)
+    int lround(double x)
 
-cdef long max_fib = 4000000
+cdef int max_fib = 4000000
 
 # The ratio of two consecutive terms in the Fibonacci sequence approaches
 # phi, the golden ratio, approximately 1.61803...
@@ -21,9 +21,9 @@ cdef double phi = (1 + sqrt(5))/2
 # Thus, we can calculate the next even Fibonacci number by multiplying
 # the previous even Fibonacci number by phi^3
 
-cdef long nextEvenFib, runningSum
+cdef int nextEvenFib, runningSum
 
-cdef long sumEvenFibs(long currentEvenFib):
+cdef int sumEvenFibs(int currentEvenFib):
     if currentEvenFib <= max_fib:
         nextEvenFib = lround(currentEvenFib * phi**3)
         runningSum = currentEvenFib + sumEvenFibs(nextEvenFib)
@@ -33,6 +33,6 @@ cdef long sumEvenFibs(long currentEvenFib):
 
 # Start summing from 2, as zero is negligible
 
-cdef long total = sumEvenFibs(2)
+cdef int total = sumEvenFibs(2)
 
 print total
